mlab-ns2: DB model from mlab-ns ported to Go.
This is a direct copy of the DB model that was used in the Python version
of mlab-ns.  There are quite a few things here that don't make sense; for
example, why would you store a port number as a string?  This can be used
to access data that is copied from the existing mlab-ns application into
mlab-ns2, but we should consider optimizations that should be made before
deploying this into production.  Now would be a good time to talk about
those.

R=bipasa, tiziana, seon.wook
CC=gavaletz, gsoc-2013
https://codereview.appspot.com/13056044