mlab-ns2: Implementing a client for rfc2617 as a http.RoundTripper.
To authenticate to apache servers that use mod_digest or mod_auth_digest
we need to implement HTTP Digest Access Authentication.  This CL covers
the most common case (MD5 and qop=auth).  The other parts of rfc2617 are
not well represented on production servers and are not needed at this
time.

R=dominic
CC=gavaletz, gsoc-2013
https://codereview.appspot.com/13386045